// frontend/src/types/api/testMailInput.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */
import type { TestMailInputHeaders } from './testMailInputHeaders';

/**
 * Sample mail data for testing a rule.
 */
export interface TestMailInput {
  sender?: string;
  sender_name?: string;
  recipient?: string;
  subject?: string;
  body?: string;
  headers?: TestMailInputHeaders;
  date?: string;
  has_attachments?: boolean;
  attachment_names?: string[];
  mail_size?: number;
  is_reply?: boolean;
  is_forwarded?: boolean;
  contact_name?: string | null;
  contact_org?: string | null;
}